refactor: split and split
